+++ lib/interfaces/ReplayMessagePayload.ts
@@ -1,0 +1,5 @@
+import { KafkaMessage } from "sinek";
+
+export interface ReplayMessagePayload {
+    message: KafkaMessage;
+}