+++ lib/interfaces/RetryMessagePayload.ts
@@ -1,0 +1,7 @@
+import { KafkaMessage } from "sinek";
+
+export interface RetryMessagePayload {
+    hookId?: string;
+    retryCount: number;
+    message: KafkaMessage;
+}